Add UnityNetBase that is used as a base class for UnityClient/Server
