+++ Assets/Exanite.Arpg/Networking/UnityNetBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 716b199d16af82d4680bb3d909086e4d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
